mcp-server/mcp-azure-sql/requirements.txt
python-dotenv
uvicorn
pydantic
fastmcp
pyodbc
azure-identity